requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.2
Pillow
django-cors-headers==4.2.0
python-decouple==3.8
setuptools